Turns out that moving POSTHyperlinkNote to a better package is not serialization-compatible.
